+++ plugin-dev/Source/Sentry/Private/Android/AndroidSentrySubsystem.h
@@ -45,6 +45,13 @@ public:
 	virtual TSharedPtr<ISentryTransactionContext> ContinueTrace(const FString& sentryTrace, const TArray<FString>& baggageHeaders) override;
 
 	virtual void HandleAssert() override;
+
+	FString TryCaptureScreenshot() const;
+
+private:
+	bool isScreenshotAttachmentEnabled = false;
+
+	FDelegateHandle OnHandleSystemErrorDelegateHandle;
 };
 
 typedef FAndroidSentrySubsystem FPlatformSentrySubsystem;